semes to be a success. Will now add the style and after that i will work on categories and search
