+++ api/v1/app/controlers/HomeController.php
@@ -3,16 +3,17 @@ require_once '/xampp/htdocs/Second_Academia_Shop/Second-Hand-shop/api/v1/app/con
 require_once '/xampp/htdocs/Second_Academia_Shop/Second-Hand-shop/api/v1/app/views/View.php';
 require_once '/xampp/htdocs/Second_Academia_Shop/Second-Hand-shop/api/v1/app/views/ViewProduct.php';
 require_once '/xampp/htdocs/Second_Academia_Shop/Second-Hand-shop/api/v1/app/modles/Product/ProductSmall.php';
+require_once '/xampp/htdocs/Second_Academia_Shop/Second-Hand-shop/api/v1/app/controlers/Header/HeaderController.php';
 class HomeController extends ABController
 {
     private $view;
     private $model;
-    private array $products;
+    private $header;
 
     public function __construct()
     {
-        $this->products = [];
         $this->view = new View;
+        $this->header = new HeaderController;
         //if ( ! class_exists('ProductSmall')) die('There is no hope!');
         $this->model = new ProductSmall("", "");
     }
@@ -25,7 +26,7 @@ class HomeController extends ABController
     }
     public function show()
     {
-        $this->view->render("/standard/header");
+        $this->header->show();
         $this->view->render("/home/home");
         $this->show_products();
         $this->view->render("/standard/footer");
